add InnerNode::NodeType()

+++ art/Node.h
@@ -5,11 +5,11 @@
 #include <memory.h>
 #include <cstddef>
 
-#define NODE4 0
-#define NODE16 1
-#define NODE48 2
-#define NODE256 3
-#define LEAFNODE 4
+#define NODE4 (int)0
+#define NODE16 (int)1
+#define NODE48 (int)2
+#define NODE256 (int)3
+#define LEAFNODE (int)4
 
 #define MAX_PREFIX_LEN 4
 #define BARRIER() __asm__ __volatile__("": : :"memory")
